Primeiro commit

+++ src/main/java/com/farmacia/demo/controllers/ProdutoController.java
@@ -2,14 +2,21 @@ package com.farmacia.demo.controllers;
 
 
 import com.farmacia.demo.dtos.ProdutoDTO;
+import com.farmacia.demo.models.Farmacia;
 import com.farmacia.demo.models.Produto;
+import com.farmacia.demo.repository.FarmaciaRepository;
 import com.farmacia.demo.repository.ProdutoRepository;
 
+import jakarta.validation.Valid;
 import org.apache.tomcat.util.http.fileupload.IOUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.MediaType;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.ModelAndView;
 
@@ -19,6 +26,8 @@ import java.io.InputStream;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.text.ParseException;
+import java.text.SimpleDateFormat;
 
 @Controller
 public class ProdutoController {
@@ -26,13 +35,21 @@ public class ProdutoController {
     @Autowired
     ProdutoRepository produtoRepository;
 
+    @Autowired
+    FarmaciaRepository farmaciaRepository;
+
     //private final String SALVAR_EM_DIR = System.getProperty("user.dir") + "\\uploads\\";
     private final String SALVAR_EM_DIR = System.getProperty("user.dir") + "\\src\\main\\resources\\static\\img\\img_produtos\\";
+
+    private Farmacia getFarmaciaLogada() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        return farmaciaRepository.findByEmail(auth.getName()).get();
+    }
 
     @GetMapping("/")
     public ModelAndView getProdutos() {
 
-        var produtos = produtoRepository.findAll();
+        var produtos = produtoRepository.findByFarmacia(getFarmaciaLogada());
         ModelAndView view = new ModelAndView("index");
         view.addObject("produtos", produtos);
 
@@ -46,17 +63,16 @@ public class ProdutoController {
     public @ResponseBody byte[] getFile(@PathVariable String imagemNome) throws IOException {
 
         Path path = Paths.get(SALVAR_EM_DIR + imagemNome);
-        if(Files.exists(path)) {
-          return Files.readAllBytes(path);
+        if (Files.exists(path)) {
+            return Files.readAllBytes(path);
         }
         return null;
     }
 
-
     @GetMapping("/buscar-produto")
-    public ModelAndView buscarProdutoPorNome(@RequestParam String nome){
+    public ModelAndView buscarProdutoPorNome(@RequestParam String nome) {
 
-        var produtos = produtoRepository.findByNomeContaining(nome);
+        var produtos = produtoRepository.findByFarmaciaAndNomeContaining(getFarmaciaLogada(), nome);
         ModelAndView view = new ModelAndView("index");
         view.addObject("produtos", produtos);
 
@@ -65,12 +81,20 @@ public class ProdutoController {
 
     @GetMapping("/cadastrar-produto")
     public String getCasdatrarProduto() {
-
         return "produto/produto_form";
     }
 
     @PostMapping("/cadastrar-produto")
-    public String postCasdatrarProduto(ProdutoDTO produtoDTO) {
+    public ModelAndView postCasdatrarProduto(@Valid ProdutoDTO produtoDTO, BindingResult result) {
+
+        ModelAndView modelAndView = new ModelAndView("produto/produto_form");
+
+        // Verifica se existe algum erro nos campos informados pelo usuário.
+        if (result.hasErrors()) {
+            modelAndView.addObject("mensagem", "Verifique os campos!");
+            modelAndView.addObject("produto", produtoDTO);
+            return modelAndView;
+        }
 
         try {
 
@@ -89,46 +113,74 @@ public class ProdutoController {
             }
 
             // Salva imagem
-            produtoDTO.getImagemFile().transferTo( new File(SALVAR_EM_DIR + nomeArquivo));
+            produtoDTO.getImagemFile().transferTo(new File(SALVAR_EM_DIR + nomeArquivo));
 
             var produto = new Produto();
             BeanUtils.copyProperties(produtoDTO, produto);
+            produto.setFarmacia(getFarmaciaLogada());
+            produto.setDataDeValidade(new SimpleDateFormat("yyyy-MM-dd").parse(produtoDTO.getDataDeValidade()));
             produto.setNomeImagem(nomeArquivo);
             produtoRepository.save(produto);
+            modelAndView.addObject("mensagem", "Produto salvo!");
 
         } catch (IOException e) {
             System.out.println("ARQUIVO NÃO SALVO");
-            e.printStackTrace();
+            modelAndView.addObject("mensagem", "Erro ao salvar produto!");
+        } catch (ParseException e) {
+            modelAndView.addObject("mensagem", "Erro na data de validade fornecida!");
+            throw new RuntimeException(e);
         }
 
-        return "redirect:/";
+        return modelAndView;
 
     }
 
     @GetMapping("/atualizar-produto/{id}")
     public ModelAndView getAtualizarProduto(@PathVariable Long id) {
 
-        var produto = produtoRepository.findById(id);
-        var view = new ModelAndView("produto/produto_form");
-        view.addObject("produto", produto.get());
+        ModelAndView modelAndView;
+        var produto = produtoRepository.findByFarmaciaAndId(getFarmaciaLogada(), id);
 
-        return view;
+        if (produto.isPresent()) {
+            modelAndView = new ModelAndView("produto/produto_form");
+            modelAndView.addObject("produto", produto.get());
+        } else {
+            modelAndView = new ModelAndView("erro404");
+        }
+
+        return modelAndView;
 
     }
 
     @PostMapping("/atualizar-produto/{id}")
-    public String postAtualizarProduto(@PathVariable Long id, ProdutoDTO produtoDTO) {
+    public ModelAndView postAtualizarProduto(@PathVariable Long id, @Valid ProdutoDTO produtoDTO, BindingResult result) {
+
+        ModelAndView modelAndView;
+        var produto = produtoRepository.findByFarmaciaAndId(getFarmaciaLogada(), id);
+
+        if (produto.isPresent()) {
+            modelAndView = new ModelAndView("produto/produto_form");
+        } else {
+            modelAndView = new ModelAndView("erro404");
+            return modelAndView;
+        }
+
+        // Verifica se existe algum erro nos campos informados pelo usuário.
+        if (result.hasErrors()) {
+            modelAndView.addObject("mensagem", "Verifique os campos!");
+            BeanUtils.copyProperties(produtoDTO, produto.get());
+            modelAndView.addObject("produto", produto.get());
+            return modelAndView;
+        }
 
         try {
-
-            var produto = produtoRepository.findById(id).get();
 
             // Se for atualizar a imagem
             if (!produtoDTO.getImagemFile().isEmpty()) {
 
                 // Deleta imagem antiga
-                Path pathDelete = Paths.get(SALVAR_EM_DIR + produto.getNomeImagem());
-                if(Files.exists(pathDelete))
+                Path pathDelete = Paths.get(SALVAR_EM_DIR + produto.get().getNomeImagem());
+                if (Files.exists(pathDelete))
                     Files.delete(pathDelete);
 
                 // Verifica se arquivo já existe e cria um novo nome caso já exista
@@ -145,22 +197,28 @@ public class ProdutoController {
                 }
 
                 // Salva a nova imagem
-                produtoDTO.getImagemFile().transferTo( new File(SALVAR_EM_DIR + nomeArquivo));
-
-                produto.setNomeImagem(nomeArquivo);
+                produtoDTO.getImagemFile().transferTo(new File(SALVAR_EM_DIR + nomeArquivo));
+                produto.get().setNomeImagem(nomeArquivo);
 
             }
 
-            BeanUtils.copyProperties(produtoDTO, produto);
-            produto.setId(id);
-            produtoRepository.save(produto);
+            BeanUtils.copyProperties(produtoDTO, produto.get());
+            produto.get().setId(id);
+            produto.get().setDataDeValidade(new SimpleDateFormat("yyyy-MM-dd").parse(produtoDTO.getDataDeValidade()));
+            produtoRepository.save(produto.get());
+            modelAndView.addObject("mensagem", "Produto atualizado!");
+
 
         } catch (IOException e) {
             System.out.println("ARQUIVO NÃO SALVO");
-            e.printStackTrace();
+            modelAndView.addObject("mensagem", "Erro ao atualizar produto!");
+        }catch (ParseException e) {
+            modelAndView.addObject("mensagem", "Erro na data de validade fornecida!");
+            throw new RuntimeException(e);
         }
 
-        return "redirect:/";
+        modelAndView.addObject("produto", produto.get());
+        return modelAndView;
 
     }
 
@@ -169,16 +227,16 @@ public class ProdutoController {
 
         try {
 
-            var produto = produtoRepository.findById(id);
+            var produto = produtoRepository.findByFarmaciaAndId(getFarmaciaLogada(), id);
             if (!produto.isPresent()) {
-                return "redirect:/?status=ERRO";
+                return "erro404";
             }
 
             Path path = Paths.get(SALVAR_EM_DIR + produto.get().getNomeImagem());
-            if(Files.exists(path))
+            if (Files.exists(path))
                 Files.delete(path);
 
-            produtoRepository.deleteById(id);
+            produtoRepository.delete(produto.get());
 
             return "redirect:/?status=OK";
